Added much more functionality to fzf.zsh
Added f, c, g, and s functions
f -> searches through select project directories for files and directories.
Opens the selection in a vim buffer
c -> searches the cwd and opens the file in an editor
g -> uses ripgrep to search the cwd for a string and goes to the
location of the string once selected.
s -> searches select folders for a given directory and cds to it once
selected